% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordvectors.R
\name{analogy}
\alias{analogy}
\alias{analogy.wordvectors}
\alias{analogy_words}
\alias{analogy_words.wordvectors}
\title{Analogy}
\usage{
analogy(model, pos, neg, n = 5L)

\method{analogy}{wordvectors}(model, pos, neg, n = 5L)

analogy_words(model, pos, neg, n = 5L)

\method{analogy_words}{wordvectors}(model, pos, neg, n = 5L)
}
\arguments{
\item{model}{A model as returned by \code{\link{word_vectors}}.}

\item{pos, neg}{Positive and negative words.}

\item{n}{Number of neightbours to return.}
}
\description{
Compute the analogy similarity between two lists of words. The positions
and the similarity values of the top \code{n} similar words will be returned.
For example: "king - man + woman =~ queen"
}
\examples{
\dontrun{
# setup word2vec Julia dependency
setup_word2vec()

# sample corpus
data("macbeth", package = "word2vec.jlr")

# train model
model_path <- word2vec(macbeth)

# get word vectors
model <- word_vectors(model_path)

# "king - man + woman =~ queen"
analogy_words(model, pos = list("king"), neg = list("malcolme"), 8L)
}

}
