% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordclusters.R
\name{as_word2cluster}
\alias{as_word2cluster}
\title{As Word2cluster}
\usage{
as_word2cluster(file)
}
\arguments{
\item{file}{Path to model file.}
}
\description{
Import a previously trained word2cluster model.
}
\examples{
\dontrun{
model_path <- as_word2cluster("path/to/file.txt")
}

}
