% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec.R
\name{as_word2vec}
\alias{as_word2vec}
\title{As model}
\usage{
as_word2vec(file)
}
\arguments{
\item{file}{Path to model file.}
}
\description{
Import a previously trained word2vec model.
}
\examples{
\dontrun{
model_path <- as_word2vec("path/to/file.txt")
}

}
