% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordclusters.R
\name{clusters}
\alias{clusters}
\alias{clusters.wordclusters}
\title{Clusters}
\usage{
clusters(model)

\method{clusters}{wordclusters}(model)
}
\arguments{
\item{model}{A model as returned by \code{\link{word_clusters}}.}
}
\description{
Return all the clusters.
}
\examples{
\dontrun{
# setup word2vec Julia dependency
setup_word2vec()

# sample corpus
data("macbeth", package = "word2vec.r")

# train model
model_path <- word2clusters(macbeth, classes = 50L)

# get word vectors
model <- word_clusters(model_path)

# get cluster
clusters(model)
}

}
