% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordvectors.R
\name{cosine}
\alias{cosine}
\alias{cosine.wordvectors}
\title{Cosine}
\usage{
cosine(model, word, n = 10L)

\method{cosine}{wordvectors}(model, word, n = 10L)
}
\arguments{
\item{model}{A model as returned by \code{\link{word_vectors}}.}

\item{word}{The word to extract.}

\item{n}{Number of neightbours to return.}
}
\value{
A \link[tibble]{tibble} of word \code{index} and their \code{cosine}.
}
\description{
Return the position of \code{n} (by default \code{n = 10}) neighbors 
of \code{word} and their cosine similarities.
}
\examples{
\dontrun{
# setup word2vec Julia dependency
setup_word2vec()

# sample corpus
data("macbeth", package = "word2vec.r")

# train model
model_path <- word2vec(macbeth)

# get word vectors
model <- word_vectors(model_path)

# neighbours of macbeth and their cosine
cosine(model, "macbeth", 20L)
}

}
