% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordvectors.R
\name{cosine_similar_words}
\alias{cosine_similar_words}
\alias{cosine_similar_words.wordvectors}
\title{Cosine Similar Words}
\usage{
cosine_similar_words(model, word, n = 10L)

\method{cosine_similar_words}{wordvectors}(model, word, n = 10L)
}
\arguments{
\item{model}{A model as returned by \code{\link{word_vectors}}.}

\item{word}{The word to extract.}

\item{n}{Number of neightbours to return.}
}
\description{
Return the top \code{n} (by default \code{n = 10}) words 
similar to \code{word}.
}
\examples{
\dontrun{
# setup word2vec Julia dependency
setup_word2vec()

# sample corpus
data("macbeth", package = "word2vec.r")

# train model
model_path <- word2vec(macbeth)

# get word vectors
model <- word_vectors(model_path)

# words similar to macbeth
cosine_similar_words(model, "macbeth", 20L)
}

}
