% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordclusters.R
\name{get_cluster}
\alias{get_cluster}
\alias{get_cluster.wordclusters}
\title{Get Cluster}
\usage{
get_cluster(model, word)

\method{get_cluster}{wordclusters}(model, word)
}
\arguments{
\item{model}{A model as returned by \code{\link{word_clusters}}.}

\item{word}{Word to retrieve the cluster.}
}
\description{
Return the cluster number for a word in the vocabulary.
}
\examples{
\dontrun{
# setup word2vec Julia dependency
setup_word2vec()

# sample corpus
data("macbeth", package = "word2vec.r")

# train model
model_path <- word2clusters(macbeth, classes = 50L)

# get word vectors
model <- word_clusters(model_path)

# get cluster
get_cluster(model, "king")
get_cluster(model, "macbeth")
}

}
