% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec.R
\name{get_vector}
\alias{get_vector}
\alias{get_vector.wordvectors}
\title{Get Vector}
\usage{
get_vector(model, word)

\method{get_vector}{wordvectors}(model, word)
}
\arguments{
\item{model}{A model as returned by \code{\link{word_vectors}}.}

\item{word}{The word to extract.}
}
\description{
Extract a specific word vector.
}
\examples{
\dontrun{
# setup word2vec Julia dependency
setup_word2vec()

# sample corpus
data("macbeth", package = "word2vec.jlr")

# train model
model_path <- word2vec(macbeth)

# get word vectors
model <- word_vectors(model_path)

# get vector of specific word
get_vector(model, "macbeth")
}

}
