% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordclusters.R
\name{get_words}
\alias{get_words}
\alias{get_words.wordclusters}
\title{Get words}
\usage{
get_words(model, cluster = 0L)

\method{get_words}{wordclusters}(model, cluster = 0L)
}
\arguments{
\item{model}{A model as returned by \code{\link{word_clusters}}.}

\item{cluster}{Cluster number as integral.}
}
\description{
Return all the words given a cluster.
}
\examples{
\dontrun{
# setup word2vec Julia dependency
setup_word2vec()

# sample corpus
data("macbeth", package = "word2vec.r")

# train model
model_path <- word2clusters(macbeth, classes = 25L)

# get word vectors
model <- word_clusters(model_path)

# get cluster
get_words(model, 2L)
}

}
