% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordvectors.R
\name{in_vocabulary}
\alias{in_vocabulary}
\alias{in_vocabulary.wordvectors}
\title{In Vocabulary}
\usage{
in_vocabulary(model, word)

\method{in_vocabulary}{wordvectors}(model, word)
}
\arguments{
\item{model}{A model as returned by \code{\link{word_vectors}}.}

\item{word}{The word to extract.}
}
\description{
Return whether a specific word is in the vocabulary.
}
\examples{
\dontrun{
# setup word2vec Julia dependency
setup_word2vec()

# sample corpus
data("macbeth", package = "word2vec.r")

# train model
model_path <- word2vec(macbeth)

# get word vectors
model <- word_vectors(model_path)

# check if James I is mentioned
in_vocabulary(model, "james")
}

}
