% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{macbeth}
\alias{macbeth}
\title{Macbeth}
\format{A character string (\code{vector} of length 1).}
\source{
\url{https://www.gutenberg.org/ebooks/2264}
}
\usage{
macbeth
}
\description{
Macbeth by William Shakespeare.
}
\keyword{datasets}
