% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_word2vec}
\alias{setup_word2vec}
\title{Setup a session}
\usage{
setup_word2vec(...)
}
\arguments{
\item{...}{Arguments passed to \link[JuliaCall]{julia_setup}.}
}
\description{
Setup a session, installs Word2Vec Juia dependency if needed.
}
