% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordvectors.R
\name{similarity}
\alias{similarity}
\alias{similarity.wordvectors}
\title{Similarity}
\usage{
similarity(model, word1, word2)

\method{similarity}{wordvectors}(model, word1, word2)
}
\arguments{
\item{model}{A model as returned by \code{\link{word_vectors}}.}

\item{word1, word2}{Words to compare.}
}
\description{
Return the cosine similarity value between two words \code{word1} and \code{word2}.
}
\examples{
\dontrun{
# setup word2vec Julia dependency
setup_word2vec()

# sample corpus
data("macbeth", package = "word2vec.r")

# train model
model_path <- word2vec(macbeth)

# get word vectors
model <- word_vectors(model_path)

# similarity b/w macbeth and wife
similarity(model, "macbeth", "king")
}

}
