% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordvectors.R
\name{size}
\alias{size}
\alias{size.wordvectors}
\title{Size}
\usage{
size(model)

\method{size}{wordvectors}(model)
}
\arguments{
\item{model}{A model as returned by \code{\link{word_vectors}}.}
}
\description{
Return the word vector length and the number of words.
}
\examples{
\dontrun{
# setup word2vec Julia dependency
setup_word2vec()

# sample corpus
data("macbeth", package = "word2vec.r")

# train model
model_path <- word2vec(macbeth)

# get word vectors
model <- word_vectors(model_path)

# check size of model
size(model)
}

}
