% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec.R
\name{word2vec}
\alias{word2vec}
\title{Word2Vec}
\usage{
word2vec(train, output = NULL, size = 100L, window = 5L,
  sample = 0.001, hs = 0L, negative = 5L, threads = 12L,
  iter = 5L, min_count = 5L, alpha = 0.025, debug = 2L,
  binary = 1L, cbow = 1L, verbose = TRUE, ...)
}
\description{
Train Word2Vec model.
}
