% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordclusters.R
\name{word_clusters}
\alias{word_clusters}
\alias{word_clusters.word2clusters}
\title{Word Clusters}
\usage{
word_clusters(file)

\method{word_clusters}{word2clusters}(file)
}
\arguments{
\item{file}{Path to model, output of \code{\link{word2clusters}} or \code{\link{as_word2clusters}}.}
}
\description{
Generate a word clusters from the model file.
}
\examples{
\dontrun{
# setup word2vec Julia dependency
setup_word2vec()

# sample corpus
data("macbeth", package = "word2vec.r")

# train model
model_path <- word2clusters(macbeth, classes = 50L)

# get word vectors
model <- word_clusters(model_path)
}

}
\seealso{
\code{\link{word2clusters}}
}
