% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word2vec.R
\name{word_vectors}
\alias{word_vectors}
\alias{word_vectors.word2vec}
\title{Word Vectors}
\usage{
word_vectors(file)

\method{word_vectors}{word2vec}(file)
}
\arguments{
\item{file}{Path to odel, output of \code{\link{word2vec}} or \code{\link{as_word2vec}}.}
}
\description{
Import Word vectors.
}
\examples{
\dontrun{
# sample corpus
corpus <- paste("Julia has foreign function interfaces for C/Fortran,", 
  "C++, Python, R, Java, and many other languages.")

# train model
model_path <- word2vec(corpus)

# get word vectors
word_vectors(model_path)
}

}
\seealso{
\code{\link{word2vec}}
}
