% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordclusters.R
\docType{data}
\name{WordClusters}
\alias{WordClusters}
\title{Word Clusters Reference Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WordClusters
}
\description{
Convert the output of \code{word_clusters} into a convenient reference class.
}
\section{Methods}{
 
All methods applicable to objects of class \code{wordclusters} as returned 
by \code{\link{word_clusters}} are valid here. See their respective man pages
for documentation on their arguments and return values.

\itemize{
  \item{\code{\link{vocabulary}}}
  \item{\code{\link{in_vocabulary}}}
  \item{\code{\link{index}}}
  \item{\code{\link{get_cluster}}}
  \item{\code{\link{clusters}}}
  \item{\code{\link{get_words}}}
}
}

\examples{
\dontrun{
# setup word2vec Julia dependency
setup_word2vec()

# sample corpus
data("macbeth", package = "word2vec.r")

# train model
model_path <- word2clusters(macbeth, classes = 15L)

# get word vectors
model <- word_clusters(model_path)
wc <- WordClusters$new(model)
wc$get_words(4L)
wc$in_vocabulary("cake")
}

}
\keyword{datasets}
