% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordvectors.R
\docType{data}
\name{WordVectors}
\alias{WordVectors}
\title{Word Vectors Reference Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WordVectors
}
\description{
Convert the output of \code{word_vectors} into a convenient reference class.
}
\section{Methods}{
 
All methods applicable to objects of class \code{wordvectors} as returned 
by \code{\link{word_vectors}} are valid here. See their respective man pages
for documentation on their arguments and return values.

\itemize{
  \item{\code{\link{get_vector}}}
  \item{\code{\link{vocabulary}}}
  \item{\code{\link{in_vocabulary}}}
  \item{\code{\link{size}}}
  \item{\code{\link{index}}}
  \item{\code{\link{cosine}}}
  \item{\code{\link{cosine_similar_words}}}
  \item{\code{\link{similarity}}}
  \item{\code{\link{analogy}}}
  \item{\code{\link{analogy_words}}}
}
}

\examples{
\dontrun{
# setup word2vec Julia dependency
setup_word2vec()

# sample corpus
data("macbeth", package = "word2vec.jlr")

# train model
model_path <- word2vec(macbeth)

# get word vectors
model <- word_vectors(model_path)
wv <- WordVectors$new(model)
wv$get_vector("king")
wv$in_vocabulary("cake")
}

}
\keyword{datasets}
