% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordvectors.R
\name{vocabulary}
\alias{vocabulary}
\alias{vocabulary.wordvectors}
\alias{vocabulary.wordclusters}
\title{Get Vocabulary}
\usage{
vocabulary(model)

\method{vocabulary}{wordvectors}(model)

\method{vocabulary}{wordclusters}(model)
}
\arguments{
\item{model}{A model as returned by \code{\link{word_vectors}} or
\code{\link{word_clusters}}.}
}
\description{
Return the vocabulary as a vector of words.
}
\examples{
\dontrun{
# setup word2vec Julia dependency
setup_word2vec()

# sample corpus
data("macbeth", package = "word2vec.r")

# train model
model_path <- word2vec(macbeth)

# get word vectors
model <- word_vectors(model_path)

# get vocabulary
vocab <- vocabulary(model)
head(vocab)
}

}
