% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordvectors.R
\name{word_vectors}
\alias{word_vectors}
\alias{word_vectors.word2vec}
\title{Word Vectors}
\usage{
word_vectors(file)

\method{word_vectors}{word2vec}(file)
}
\arguments{
\item{file}{Path to odel, output of \code{\link{word2vec}} or \code{\link{as_word2vec}}.}
}
\description{
Import Word vectors.
}
\examples{
\dontrun{
# setup word2vec Julia dependency
setup_word2vec()

# sample corpus
data("macbeth", package = "word2vec.jlr")

# train model
model_path <- word2vec(macbeth)

# get word vectors
model <- word_vectors(model_path)
}

}
\seealso{
\code{\link{word2vec}}
}
